requirements.txt
requests>=2.28.0
python-dateutil>=2.8.2
tzdata>=2022.1
python-dotenv>=1.0.0
# Google ADK dependencies
google-generativeai>=0.3.0
protobuf>=4.22.3 